requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==23.2.0
azure-core==1.32.0
azure-identity==1.19.0
billiard==4.2.1
boto3==1.34.153
botocore==1.34.153
cachetools==5.5.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.1
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.1
dataclasses-json==0.6.7
distro==1.9.0
Django==4.1
django-cors-headers==4.4.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
faiss-cpu
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
greenlet==3.0.3
gspread==6.1.4
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.25.2
idna==3.7
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.4.2
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.28
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langsmith==0.1.96
marshmallow==3.21.3
msal==1.31.1
msal-extensions==1.2.0
multidict==6.0.5
mypy-extensions==1.0.0
neo4j==5.23.0
nltk==3.9.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.38.0
openpyxl==3.1.5
orjson==3.10.6
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
portalocker==2.10.1
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.3
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.10.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.2.0
pypdf==4.3.1
pypdfium2==4.30.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.2.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.5
schedule==1.2.2
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.31
sqlparse==0.5.1
tablib==3.6.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.1
tqdm==4.66.4
transformers==4.45.2
trio==0.27.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
